Add some data and (very) basic routes
The data added is a huge JSON file containing all the references that appear
in research guides, along with backlinks. For testing purposes a smaller 'sample'
file has also been added.

This provides a home route (that returns a home page) and a very simple
implementation of a `/guides/<ref>` route that returns a JSON object, provided
a match is found at the root of the JSON.